Convert extension template and Makefile to asciidoctor
Extension HTML docs now use the same toolchain and CSS as the OpenCL API
HTML doc.

Something people might care about is that the individual extension
HTML/PDF now don't have a TOC. This could probably be fixed with some
fairly straightforward conditionals in the documents to check whether
they are being included in the unified extensions doc or built
standalone.

This works as is, but it would probably make more sense to use the
parent Makefile to generate these documents as well as the API, C++,
etc. specs.

Closes internal issue 162.