Add static refpages. All of these are for extensions, but not easily extractable from the OpenCL Extensions document at present. There are a few minor tweaks to the spec sources and changes to the Makefile to incorporate the static files into the refpage build, and create the refpage navigation and sidebar indices.
The static pages are imported largely unchanged from the Docbook
versions, aside from reformatting to match the other pages. This means
that there are a lot of out-of-date references to explicit section and
table numbers in other documents, and that they have incomplete coverage
of extensions (in particular, the EXTENSION page is not accurate about
what extensions may be enabled).

However, once we've merged this, I'll be able to replace the old
refpages and we can move forward from there as the WG needs to.