WIP: fix refpage links not opening
This adds '^' to generated asciidoctor link: text, causing the resulting
HTML to open external hyperlinks in a new page. While that works, there
are some related issues I'm exploring with landing on the wrong anchor
in the opened document when it's a CL specification, so marking this WIP
for the moment.