Clarify default value of program-scope/static variables
This patch clarifies that OpenCL program-scope/static variables in the
global address space have an appropriate null value by default,
similar to objects with static storage duration in C99 6.7.8p10.

The previous wording was not clear regarding the default value (is the
variable uninitialized?) when no constant expression is given. This
clarifies the intent.