Allow use of CL_DEVICE_COMPILER_AVAILABLE for IL
Currently how the use of IL programs should behave when
`CL_DEVICE_COMPILER_AVAILABLE` is `CL_FALSE` is underspecified.  This change
allows `CL_DEVICE_COMPILER_AVAILABLE` to be returned by entry points that
require use of a compiler.

`clBuildProgram` could already return `CL_DEVICE_COMPILER_AVAILABLE` for OpenCL
source, so extend that to IL.  `clCompileProgram` can already return
`CL_DEVICE_COMPILER_AVAILABLE` for any input type.

I've also added the `CL_DEVICE_COMPILER_AVAILABLE` to
`clSetProgramSpecializationConstant` because for it to be able to check the
validity of the provided `spec_id` it must be able to parse the IL program.

Fixes #438