remove the cl_image_desc ifdef from the XML file
The ifdef is currently rendering poorly in the generated spec
snippet, and what is in the XML file does not match what is
in the headers currently anyhow.

We will need to resolve this if and only if we decide to generate
actual headers for cl_image_desc from the XML file.