Improved description for pipe specifier.
Pipe specifies a very distinct type from its
element and therefore a better wording describing
it should be 'type specifier'.

Note that pipe can not be a type qualifier because
in C-based languages a type qualifier is an optional
property that doesn't modify representation of the
type (i.e. casting away qualifiers is valid in certain
circumstances). OpenCL spec never allowed to cast away
pipe and doing so won't result in anythinig useful
unless there is some knowledge of its implementation.