Clarify ballot ops for remainder subgroups (#689)
As discussed in issue #626, the wording used to describe the
sub_group_ballot_bit_count() and sub_group_ballot_find_msb() operations
included the phrase "only considering the bits in _value_ that represent
predicate values from all work items in the subgroup", but "remainder"
subgroups were not considered when arriving at this wording.

In Vulkan, the subgroup size is defined in a way that is equivalent to
the definition of the maximum subgroup size in OpenCL.  Therefore, in
Vulkan, "remainder" subgroups start with work-items that are initially
inactive, but are still considered work-items within the subgroup.

The ballot operations in OpenCL were intended to be equivalent to those
in Vulkan.  Therefore, for a "remainder" subgroup, predicate bits
that would correspond to non-existent work-items that would be
considered initially inactive in Vulkan should still contribute to
ballot operations that concern predicate bits and are always required
to be well-defined.

The sub_group_ballot_bit_count() and sub_group_ballot_find_msb()
operations are affected by this.

Due to the manner of their operation, the sub_group_ballot_find_lsb(),
sub_group_ballot_inclusive_scan() and sub_group_ballot_exclusive_scan()
operations are not affected.  However, as sub_group_ballot_find_lsb()
is described using the same wording, and altering the wording does not
affect its semantics, it has been updated for consistency.