Remove command-buffer Invalid state
It was discovered during cl_khr_command_buffer layered extension review
that the
[Invalid command-buffer state](https://registry.khronos.org/OpenCL/specs/3.0-unified/html/OpenCL_Ext.html#_add_new_section_section_5_x_1_command_buffer_lifecycle)
is undesirable.

An approach more in-keeping with the OpenCL philosophy is for an OpenCL
object to be able to update the reference count of the OpenCL objects it
uses. Keeping those objects alive for its lifetime, rather than having a
specific validity check. This change specifies that command-buffers
match that behaviour. Any validity checking further than that may be
expensive, and should not be mandated.

There is also a clarification in this change that interleaving queue
submissions with command recording to the same queue is valid.