Remove command-buffer Invalid state (#885)
* Remove command-buffer Invalid state

It was discovered during cl_khr_command_buffer layered extension review
that the
[Invalid command-buffer state](https://registry.khronos.org/OpenCL/specs/3.0-unified/html/OpenCL_Ext.html#_add_new_section_section_5_x_1_command_buffer_lifecycle)
is undesirable.

An approach more in-keeping with the OpenCL philosophy is for an OpenCL
object to be able to update the reference count of the OpenCL objects it
uses. Keeping those objects alive for its lifetime, rather than having a
specific validity check. This change specifies that command-buffers
match that behaviour. Any validity checking further than that may be
expensive, and should not be mandated.

There is also a clarification in this change that interleaving queue
submissions with command recording to the same queue is valid.

* Warning about clSetKernelArg

Add non-normative text warning that objects used
as arguments to kernels recorded to a command-buffer do
not have their reference count updated.

* American English spelling of "behavior"

Change "behaviour" to "behavior" in command-buffer
related specs.

* Use term "attached"

The term "attached" is more consistent with the existing OpenCL
spec with regards to how reference counting is specified.

* Expand kernel argument note about safe usage

Elaborate on the note about ref counting not being
done on kernel arguments, with recommendations for users
not to free objects used as kernel arguments until
the command-buffer is deleted.

* Move mutable-dispatch note to own spec

Based on working-group feedback that the information
about mutable-dispatch safe usage shouldn't live in the
base specification, but instead be linked to.