Remove CRLF line endings from two core spec files (#1079)
All other text files in the repository use LF line endings.  The presence of
the carriage returns in the files means that on macOS the shell code that
automatically builds up the Makefile dependencies from the include lines in
Asciidoc was not working as xargs did not behave as intended.

Note that various Intel vendor extensions also use CRLF line endings but this
PR does not touch them, they do not contain includes so do not actually cause
any problems.