tidy up cl_khr_create_command_queue
Removes unneeded special-case text regarding the type accepted for
CL_QUEUE_PROPERTIES.  Calling this a cl_bitfield was a bug all
along, and using cl_command_queue_properties is fine because it is
a bitfield anyhow.

Generalizes some text so it is not specific to OpenCL 2.x devices.

Removes a TODO - no changes required.