cl_ext_buffer_device_address
The basic cl_mem buffer API doesn't enable access to the underlying raw
pointers in the device memory, preventing its use in host side
data structures that need pointer references to objects.
This API adds a minimal increment on top of cl_mem that provides such
capabilities.

The version 0.1.0 is implemented in PoCL and rusticl for prototyping,
but everything's still up for discussion. chipStar is the first client
that uses the API.