Add revision to XML extensions (#1161)
This extension adds the revision field to the
XML tag for extensions. This allows a version
macro to be generated with:

* https://github.com/KhronosGroup/OpenCL-Headers/pull/251
* https://github.com/KhronosGroup/OpenCL-Headers/issues/248

KHR extensions are given a revision based on the semantic
version of the spec. However other extensions don't use
semantic versioning, and so are given a placeholder `0.0.0`
value until they can be updated by the owner.

The XML schema is also updated to make the revision field
mandatory in the XML entry for extensions and the
existence of the macro this enables is advertised to users
in the spec.