Refactor command-buffer queue compatability
As proposed in https://github.com/KhronosGroup/OpenCL-Docs/issues/1142
the PR changes the semantics of the command-queues parameters used for
command-buffer creation and enqueue.

The queues used on command-buffer creation now only inform the
device and dependencies of commands, rather than restricting the
properties set on the queues used for command-buffer enqueue.

This is based ontop on the change in https://github.com/KhronosGroup/OpenCL-Docs/pull/850
to add supported queue property semantics.