cl_ext_buffer_device_address (#1159)
* cl_ext_buffer_device_address

The basic cl_mem buffer API doesn't enable access to the underlying raw
pointers in the device memory, preventing its use in host side
data structures that need pointer references to objects. This API
adds a minimal increment on top of cl_mem that provides such
capabilities.

* BDA: Removed CL_MEM_DEVICE_SHARED_ADDRESS_EXT as unneeded.

Also made the enums globally unique.

* cl_ext_buffer_device_address to 1.0.0

The only content addition since the previous version is

"If the device supports SVM and {clCreateBufferWithProperties} is
 called with a pointer returned by {clSVMAlloc} as its _host_ptr_
 argument, and {CL_MEM_USE_HOST_PTR} is set in its _flags_ argument,
 the device-side address is guaranteed to match the _host_ptr."

* cl_ext_buffer_device_address: Revision 1.0.1

* Made it explicit that passing illegal pointers is legal as long
as they are not referenced.
* Removed CL_INVALID_ARG_VALUE as a possible error in
clSetKernelArgDevicePointerEXT() as there are no illegal pointer
cases when calling this function. Return CL_INVALID_OPERATION for
clGetMemObjectInfo() if the pointer is not a buffer device
pointer.
* clSetKernelExecInfo() and clSetKernelArgDevicePointerEXT() now
only error out if no devices in the context associated with kernel
support device pointers.

* cl_ext_buffer_device_address: Revision 1.0.2

Converted the clSetKernelArgDevicePointerEXT() address parameter to
a value instead of a pointer to the value.