Fix up command-buffer enqueue error after PR-1382 (#1471)
https://github.com/KhronosGroup/OpenCL-Docs/pull/1382 added support
for deferring setting arguments on a kernel command which is updatable.
To achieve this we added an extra command-buffer state, "finalized"
which is entered when a command-buffer has been finalized but doesn't
yet have all it's arguments set.

However, if a user tries to enqueue a command-buffer in this state
it shouldn't be valid. Therefore update our current wording about
when an command-buffer can be enqueued to say the specifically the executable
state, when all arguments are available. Rather than defined as after the
finalization operation, which could result in the command-buffer being in
either the finalization or executable state.

We already have a NOTE above to this effect, which I've removed as it
felt like it didn't add any extra value above the error definition, but
can added back if folks think it's worthwhile.